CU-31v9g3c_EXCEPTION-HANDLING_Avinash-Vijayvargiya (#27)
Added exception handling

+++ src/main/java/com/duosec/duosecbackend/service/DashboardService.java
@@ -3,7 +3,11 @@ package com.duosec.duosecbackend.service;
 import com.duosec.duosecbackend.dao.AuthModel;
 import com.duosec.duosecbackend.dao.DashboardModel;
 import com.duosec.duosecbackend.dto.*;
+import com.duosec.duosecbackend.exception.DataException;
+import com.duosec.duosecbackend.exception.EmptyDataException;
+import com.duosec.duosecbackend.exception.NullDataException;
 import com.duosec.duosecbackend.model.CompanyEmployee;
+import com.duosec.duosecbackend.utils.ExtensionFunction;
 import org.duosec.backendlibrary.SecretGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -30,7 +34,6 @@ public class DashboardService {
 
     @Autowired
     private AuthModel authModel;
-
 
     public String addEmployee(AddEmployeeData addEmployeeData) {
         byte[] secret = SecretGenerator.generate();
@@ -60,24 +63,48 @@ public class DashboardService {
         return "Data Saved";
     }
 
-    public String deleteEmployee(DeleteEmployeeData deleteEmployeeData) {
-        dashboardModel.deleteByEmployeeIdAAndCompanyUniqueId(deleteEmployeeData.getEmployeeId(), deleteEmployeeData.getCompanyUniqueId());
-        return "Data Deleted";
+    public String deleteEmployee(DeleteEmployeeData deleteEmployeeData) throws NullDataException, EmptyDataException, DataException {
+        ExtensionFunction extensionFunction = new ExtensionFunction();
+        if (extensionFunction.isNull(deleteEmployeeData.getEmployeeId()) || extensionFunction.isNull(deleteEmployeeData.getCompanyUniqueId()))
+            throw new EmptyDataException("Delete Employee Can't be Null");
+        if (deleteEmployeeData.getEmployeeId().isEmpty() || deleteEmployeeData.getCompanyUniqueId().isEmpty())
+            throw new EmptyDataException("Delete Employee Can't be Empty");
+
+        try {
+            dashboardModel.deleteByEmployeeIdAAndCompanyUniqueId(deleteEmployeeData.getEmployeeId(), deleteEmployeeData.getCompanyUniqueId());
+            return "Data Deleted";
+        } catch (DataException dataException) {
+            throw new DataException("Data Not Deleted");
+        }
     }
 
-    public String deleteEmployee(DeleteEmployeeDataAPI deleteEmployeeDataAPI) {
-        String companyUniqueId = authModel.findByApiKey(deleteEmployeeDataAPI.getApiKey()).getCompanyUniqueId();
-        dashboardModel.deleteByEmployeeIdAAndCompanyUniqueId(deleteEmployeeDataAPI.getEmployeeId(), companyUniqueId);
-        return "Data Deleted";
+    public String deleteEmployee(DeleteEmployeeDataAPI deleteEmployeeDataAPI) throws NullDataException, EmptyDataException, DataException {
+        ExtensionFunction extensionFunction = new ExtensionFunction();
+        if (extensionFunction.isNull(deleteEmployeeDataAPI.getApiKey()) || extensionFunction.isNull(deleteEmployeeDataAPI.getEmployeeId()))
+            throw new EmptyDataException("Delete Employee Can't be Null");
+        if (deleteEmployeeDataAPI.getApiKey().isEmpty() || deleteEmployeeDataAPI.getEmployeeId().isEmpty())
+            throw new EmptyDataException("Delete Employee Can't be Empty");
+
+        try {
+            String companyUniqueId = authModel.findByApiKey(deleteEmployeeDataAPI.getApiKey()).getCompanyUniqueId();
+            dashboardModel.deleteByEmployeeIdAAndCompanyUniqueId(deleteEmployeeDataAPI.getEmployeeId(), companyUniqueId);
+            return "Data Deleted";
+        } catch (DataException dataException) {
+            throw new DataException("Data Not Saved");
+        }
     }
 
-    public Map<String, Object> getAllEmployee(String companyUniqueId, String employeeName, int page, int size, String sortBy, boolean sort) throws Exception {
+    public Map<String, Object> getAllEmployee(String companyUniqueId, String employeeName, int page, int size, String sortBy, boolean sort) throws NullDataException, EmptyDataException {
         List<DashboardAllData> dashboardAllDataList = new ArrayList<>();
         Page<CompanyEmployee> companyEmployeePage;
         Pageable paging;
 
-        if (companyUniqueId == null)
-            throw new Exception("Null Company Unique Id");
+        ExtensionFunction extensionFunction = new ExtensionFunction();
+        if (extensionFunction.isNull(companyUniqueId))
+            throw new NullDataException("Null Company Unique Id");
+
+        if (companyUniqueId.isEmpty())
+            throw new EmptyDataException("Empty Company Unique Id");
 
         if (size < 10)
             size = 10;
